requirements.txt
torchvision==0.8.1+cpu
scikit_image==0.18.2
numpy==1.19.3
torchnet==0.0.4
torch==1.7.0+cpu
matplotlib==3.3.3
fire==0.4.0
prettytable==2.0.0
Pillow==8.3.1
skimage==0.0
